Fixed sizing of new articles/captions!!!!!
When you drag to create a new article/caption, the box being drawn now
follows the mouse. The bug was that the box would seem to draw randomly
somewhere else in the iPad.

Also fixed a bug where if an article had no text, there would be a JS
error.